+++ axelor-account/src/main/java/com/axelor/apps/account/service/invoice/InvoiceServiceImpl.java
@@ -428,7 +428,6 @@ public class InvoiceServiceImpl extends InvoiceRepository implements InvoiceServ
   @Transactional(rollbackOn = {Exception.class})
   public Invoice createRefund(Invoice invoice) throws AxelorException {
 
-    log.debug("Créer un avoir pour la facture {}", new Object[] {invoice.getInvoiceId()});
     Invoice refund = new RefundInvoice(invoice).generate();
     invoice.addRefundInvoiceListItem(refund);
     invoiceRepo.save(invoice);